services/recommendation_service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
google-genai==0.8.0